requirements.txt
aiohttp[speedups] >= 3.9.3
aiohttp-retry >= 2.8.3

backoff >= 2.2.1
boto3 >= 1.26.165
click >= 8.1.7
colorama >= 0.2.5, < 0.4.7
cryptography < 47.0.0
fastapi[all] >= 0.94.0
filelock >= 3.0.0
paramiko >= 3.3.1
prettytable >= 3.9.0
psutil >= 5.9.0
py-cpuinfo >= 9.0.0
inquirerpy == 0.3.4
requests >= 2.31.0
tomli >= 2.0.1
tomlkit >= 0.12.2
tqdm-loggable >= 0.1.4
urllib3 >= 1.26.6
watchdog >= 3.0.0
